# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Claude Code Zen API
  description: |
    Comprehensive API documentation for Claude Code Zen - AI-powered development coordination platform.
    
    This API provides endpoints for:
    - System health and status monitoring
    - Swarm management and coordination
    - Task management and orchestration
    - Document and file operations
    - LLM statistics and analytics
    - Matron advisory system
    - Visionary roadmap planning
    - Workspace management
    - Real-time WebSocket communications
    
    The API follows RESTful principles and returns JSON responses.
  version: 2.0.0
  contact:
    name: Claude Code Zen Team
    url: https://github.com/zen-neural/claude-code-zen
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://fra-d1.in.centralcloud.net/api
    description: Production server
paths:
...
  /matron/consultations:
    get:
      summary: Get Consultations
      description: Get all matron advisory consultations
      operationId: getConsultations
      tags:
        - Matron Advisory
      parameters:
        - name: status
          in: query
          description: Filter consultations by status
          required: false
          schema:
            type: string
            enum: [active, completed, pending, cancelled]
        - name: domain
          in: query
          description: Filter by expertise domain
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Consultations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  consultations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Consultation'
        '500':
          description: Failed to get consultations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create Consultation
      description: Create a new matron advisory consultation
      operationId: createConsultation
      tags:
        - Matron Advisory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConsultationRequest'
            example:
              title: "Architecture Decision: Microservices vs Monolith"
              domain: "system-architecture"
              question: "Should we migrate from monolith to microservices?"
              context:
                projectType: "E-commerce Platform"
                teamSize: "8 developers"
                timeline: "6 months"
              priority: "high"
              requestedBy: "Tech Lead"
      responses:
        '201':
          description: Consultation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Consultation'
        '400':
          description: Invalid consultation data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to create consultation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /roadmap/roadmaps:
    get:
      summary: Get Roadmaps
      description: Get all roadmaps with their current status and progress
      operationId: getRoadmaps
      tags:
        - Visionary Roadmap
      parameters:
        - name: status
          in: query
          description: Filter roadmaps by status
          required: false
          schema:
            type: string
            enum: [active, planning, completed, on-hold, cancelled]
        - name: priority
          in: query
          description: Filter roadmaps by priority
          required: false
          schema:
            type: string
            enum: [low, medium, high, critical]
      responses:
        '200':
          description: Roadmaps retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  roadmaps:
                    type: array
                    items:
                      $ref: '#/components/schemas/Roadmap'
        '500':
          description: Failed to get roadmaps
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create Roadmap
      description: Create a new roadmap for strategic planning
      operationId: createRoadmap
      tags:
        - Visionary Roadmap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoadmapRequest'
            example:
              title: "AI Platform Evolution 2025"
              description: "Roadmap for advancing AI capabilities"
              vision: "Become the leading AI development platform"
              owner: "Product Team"
              startDate: "2024-01-01"
              endDate: "2025-12-31"
              priority: "critical"
              strategicThemes: ["AI Innovation", "Platform Excellence"]
              stakeholders: ["Engineering", "Product", "AI Research"]
              budget: "$2M"
      responses:
        '201':
          description: Roadmap created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Roadmap'
        '400':
          description: Invalid roadmap data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to create roadmap
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Resource not found"
        code:
          type: string
          description: Error code (optional)
          example: "NOT_FOUND"
        details:
          type: object
          description: Additional error details (optional)

    # System Status
...
    Consultation:
      type: object
      properties:
        id:
          type: string
          example: "cons-001"
        title:
          type: string
          example: "Architecture Decision: Microservices vs Monolith"
        domain:
          type: string
          example: "system-architecture"
        expert:
          type: string
          example: "AI Architecture Advisor"
        priority:
          type: string
          enum: [low, medium, high, critical]
          example: "high"
        status:
          type: string
          enum: [active, completed, pending, cancelled]
          example: "active"
        requestedBy:
          type: string
          example: "Tech Lead"
        created:
          type: string
          format: date-time
        context:
          type: object
          additionalProperties: true
          example:
            projectType: "E-commerce Platform"
            scalabilityRequirements: "High"
            teamSize: "8 developers"
            timeline: "6 months"
        question:
          type: string
          example: "Should we migrate from monolith to microservices architecture?"
        complexity:
          type: string
          enum: [low, medium, high, critical]
          example: "high"
        businessImpact:
          type: string
          enum: [low, medium, high, critical]
          example: "critical"
        stakeholders:
          type: array
          items:
            type: string
          example: ["Engineering Team", "Product Manager", "CTO"]
    CreateConsultationRequest:
      type: object
      required:
        - title
        - domain
        - question
        - requestedBy
      properties:
        title:
          type: string
          example: "Performance Optimization Strategy"
        domain:
          type: string
          example: "performance"
        question:
          type: string
          example: "How can we optimize our database queries for better performance?"
        requestedBy:
          type: string
          example: "Senior Developer"
        priority:
          type: string
          enum: [low, medium, high, critical]
          default: "medium"
          example: "high"
        context:
          type: object
          additionalProperties: true
          example:
            currentLoad: "10k requests/hour"
            databaseType: "PostgreSQL"
            queryComplexity: "High"
        complexity:
          type: string
          enum: [low, medium, high, critical]
          default: "medium"
          example: "high"
        businessImpact:
          type: string
          enum: [low, medium, high, critical]
          default: "medium"
          example: "high"
        stakeholders:
          type: array
          items:
            type: string
          example: ["Development Team", "Database Admin"]
...
    Roadmap:
      type: object
      properties:
        id:
          type: string
          example: "rm-zen-001"
        title:
          type: string
          example: "Claude Code Zen Platform Evolution 2024-2025"
        description:
          type: string
          example: "Comprehensive roadmap for advancing claude-code-zen capabilities"
        vision:
          type: string
          example: "Transform claude-code-zen into the most intelligent development platform"
        status:
          type: string
          enum: [active, planning, completed, on-hold, cancelled]
          example: "active"
        priority:
          type: string
          enum: [low, medium, high, critical]
          example: "critical"
        owner:
          type: string
          example: "Claude Code Zen Team"
        startDate:
          type: string
          format: date
          example: "2024-08-01"
        endDate:
          type: string
          format: date
          example: "2025-08-01"
        completion:
          type: number
          format: float
          minimum: 0
          maximum: 100
          example: 35
        strategicThemes:
          type: array
          items:
            type: string
          example: ["Swarm Intelligence", "Developer Experience", "Performance Optimization"]
        stakeholders:
          type: array
          items:
            type: string
          example: ["Development Team", "Claude Code Users", "AI Research"]
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
          example: "medium"
        budget:
          type: string
          example: "$500K"
        dependencies:
          type: array
          items:
            type: string
          example: ["Claude Code CLI Integration", "MCP Protocol Stability"]
        kpis:
          type: array
          items:
            type: string
          example: ["User Adoption Rate", "Task Completion Speed", "Error Reduction"]
    CreateRoadmapRequest:
      type: object
      required:
        - title
        - description
        - vision
        - owner
        - startDate
        - endDate
      properties:
        title:
          type: string
          example: "Mobile App Development 2024"
        description:
          type: string
          example: "Roadmap for mobile application development"
        vision:
          type: string
          example: "Create industry-leading mobile experiences"
        owner:
          type: string
          example: "Mobile Team"
        startDate:
          type: string
          format: date
          example: "2024-01-01"
        endDate:
          type: string
          format: date
          example: "2024-12-31"
        priority:
          type: string
          enum: [low, medium, high, critical]
          default: "medium"
          example: "high"
        strategicThemes:
          type: array
          items:
            type: string
          example: ["Mobile Innovation", "User Experience"]
        stakeholders:
          type: array
          items:
            type: string
          example: ["Mobile Team", "UX Design", "Product"]
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
          default: "medium"
          example: "low"
        budget:
          type: string
          example: "$250K"
        dependencies:
          type: array
          items:
            type: string
          example: ["Platform API", "Design System"]
        kpis:
          type: array
          items:
            type: string
          example: ["App Store Rating", "User Engagement", "Performance"]
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (if implemented)

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication (if implemented)
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...